Catch more unpredictable cases.
Previously a conditional branch could get away with being in the branch delay slot, provided that the branch condition evaluated false. ERET could also be in the delay slot. These cases are now flagged as being unpredictable. Furthermore, there is now a check on whether register rs is 31 for the BLTZAL, BGEZAL, BLTZALL and BGEZALL instructions (this is caught by the decoder).