Update run.sml to be more careful about the handling of IntInf.int.
This requires the latest version of L3.

Note: It is now possible to generate SML that can be compiled with MLton
without requiring the "-default-type intinf" flag. This doesn't speed up
emulation but it does mean that the generated code is a bit more portable. In
particular, the next version of Poly/ML (5.6.1?) is likely to switch to having
fixed-width integers by default.